9 спринт. 3 проверка

+++ src/handler/EpicIdHandler.java
@@ -1,9 +1,7 @@
 package handler;
 
 import com.sun.net.httpserver.HttpExchange;
-import http.HttpTaskServer;
 import manager.TaskManager;
-import model.Epic;
 
 import java.io.IOException;
 
@@ -32,31 +30,16 @@ public class EpicIdHandler extends Handler {
             sendResponse(httpExchange, 404, "Такой задачи не существует");
             return;
         }
-        switch (httpExchange.getRequestMethod()) {
-            case "PUT":
-                handlePutRequest(httpExchange, id);
-                break;
-            case "DELETE":
-                handleDeleteRequest(httpExchange, id);
-                break;
-            default:
-                sendResponse(httpExchange, 405, "Неизвестный метод");
+        if (httpExchange.getRequestMethod().equals("DELETE")) {
+            handleDeleteRequest(httpExchange, id);
+        } else {
+            sendResponse(httpExchange, 405, "Неизвестный метод");
+
         }
     }
 
     private void handleDeleteRequest(HttpExchange httpExchange, int id) throws IOException {
         taskManager.removeByIdTask(id);
         sendResponse(httpExchange, 201, "Успешно");
-    }
-
-    private void handlePutRequest(HttpExchange httpExchange, int id) throws IOException {
-        Epic epic = (Epic) taskManager.getByIdInside(id);
-        try {
-            taskManager.updateEpic(epic);
-            sendResponse(httpExchange, 201, "Успешно");
-        } catch (IllegalArgumentException e) {
-            String responseBody = e.getMessage();
-            sendResponse(httpExchange, 406, responseBody);
-        }
     }
 }
